% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{matrify_solutions}
\alias{matrify_solutions}
\title{Turned parsed solutions into a matrix of models (columns) by fluxes (rows)}
\usage{
matrify_solutions(
  solution,
  scale = FALSE,
  types = c("compounds", "reactions", "biomass")
)
}
\arguments{
\item{solution}{A dataframe of ensemble solutions imported from \code{parse_solutions()}}

\item{scale}{A boolean value to scale the fluxes to the maximum value}

\item{types}{A character vector of the types of solutions to include}
}
\description{
Turned parsed solutions into a matrix of models (columns) by fluxes (rows)
}
