% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinations.R
\name{ordinate_solutions}
\alias{ordinate_solutions}
\title{Ordinate solutions}
\usage{
ordinate_solutions(solutions, distance, dim = 2, quiet = F)
}
\arguments{
\item{solutions}{A data frame or ensemble of solutions}

\item{distance}{A distance matrix}

\item{dim}{The number of dimensions to ordinate}

\item{quiet}{If TRUE, suppresses output from the ordination function}
}
\description{
Ordinate solutions
}
